refactor(lotto): 구매 가격에 따른 번호 랜덤 생성 기능 추가
