+++ src/main/java/lotto/Lotto.java
@@ -18,19 +18,4 @@ public class Lotto {
             throw new IllegalArgumentException("[ERROR] 번호 6개를 입력해주세요.");
         }
     }
-
-    public void inputPurchaseAmount(){
-        System.out.println("구입금액을 입력해 주세요.");
-        int inputPurchaseAmount = Integer.parseInt(Console.readLine());
-    }
-
-    public void validateInputPurchaseAmount(int inputPurchaseAmount) {
-        if (inputPurchaseAmount % 1000 != 0){
-            throw new IllegalArgumentException("[ERROR] 1000원 단위로 입력해주세요.");
-        }
-    }
-
-    public int calculateRound(int inputPurchaseAmount) {
-        return inputPurchaseAmount / 1000;
-    }
 }
