+++ src/main/java/lotto/domain/Numbers.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import camp.nextstep.edu.missionutils.Console;
 import camp.nextstep.edu.missionutils.Randoms;
 
 import java.util.Arrays;
@@ -8,8 +9,23 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Numbers {
-    public void inputPurchaseAmount() {
+    public int inputPurchaseAmount = 0;
+    public int inputPurchaseAmount() {
+        System.out.println("구입금액을 입력해 주세요.");
+        int inputPurchaseAmount = Integer.parseInt(Console.readLine());
+        validateInputPurchaseAmount(inputPurchaseAmount);
+        return inputPurchaseAmount;
+    }
 
+    public void validateInputPurchaseAmount(int inputPurchaseAmount){
+        if (inputPurchaseAmount % 1000 != 0){
+            throw new IllegalArgumentException("[ERROR] 1000원 단위로 입력해주세요.");
+        }
+    }
+
+    public int calculateRound(int inputPurchaseAmount) {
+        int round = inputPurchaseAmount / 1000;
+        return round;
     }
 
     public void createLottoNumbers() {
